requirements.txt
tkinterdnd2
Pillow
requests
pyperclip
python-dotenv 
pyinstaller